renamed to load as submodule
